# Improve cloid parsing performance