--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,7 +14,7 @@ repository = "https://github.com/hyperliquid-dex/hyperliquid-rust-sdk"
 [dependencies]
 chrono = "0.4.26"
 env_logger = "0.10.0"
-ethers = {version = "0.17", features = ["eip712", "abigen"]}
+ethers = { version = "0.17", features = ["eip712", "abigen"] }
 futures-util = "0.3.28"
 hex = "0.4.3"
 http = "0.2.9"
@@ -22,10 +22,10 @@ lazy_static = "1.3"
 log = "0.4.19"
 rand = "0.8.5"
 reqwest = "0.11.18"
-serde = {version = "1.0.175", features = ["derive"]}
+serde = { version = "1.0.175", features = ["derive"] }
 serde_json = "1.0.103"
 rmp-serde = "1.0.0"
 thiserror = "1.0.44"
-tokio = {version = "1.29.1", features = ["full"]}
-tokio-tungstenite = {version = "0.20.0", features = ["native-tls"]}
-uuid = {version = "1.6.1", features = ["v4"]}
+tokio = { version = "1.29.1", features = ["full"] }
+tokio-tungstenite = { version = "0.20.0", features = ["native-tls"] }
+uuid = { version = "1.6.1", features = ["v4", "serde"] }