--- a/src/exchange/order.rs
+++ b/src/exchange/order.rs
@@ -1,6 +1,6 @@
 use crate::{
     errors::Error,
-    helpers::{float_to_string_for_hashing, uuid_to_hex_string},
+    helpers::{as_hex_option, float_to_string_for_hashing},
     prelude::*,
 };
 use serde::{Deserialize, Serialize};
@@ -42,8 +42,13 @@ pub struct OrderRequest {
     pub reduce_only: bool,
     #[serde(rename = "t", alias = "orderType")]
     pub order_type: Order,
-    #[serde(rename = "c", alias = "cloid", skip_serializing_if = "Option::is_none")]
-    pub cloid: Option<String>,
+    #[serde(
+        rename = "c",
+        alias = "cloid",
+        serialize_with = "as_hex_option",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub cloid: Option<Uuid>,
 }
 
 pub struct ClientLimit {
@@ -82,8 +87,6 @@ impl ClientOrderRequest {
         };
         let &asset = coin_to_asset.get(&self.asset).ok_or(Error::AssetNotFound)?;
 
-        let cloid = self.cloid.map(uuid_to_hex_string);
-
         Ok(OrderRequest {
             asset,
             is_buy: self.is_buy,
@@ -91,7 +94,7 @@ impl ClientOrderRequest {
             limit_px: float_to_string_for_hashing(self.limit_px),
             sz: float_to_string_for_hashing(self.sz),
             order_type,
-            cloid,
+            cloid: self.cloid,
         })
     }
 }