--- a/src/helpers.rs
+++ b/src/helpers.rs
@@ -3,6 +3,7 @@ use chrono::prelude::Utc;
 use lazy_static::lazy_static;
 use log::info;
 use rand::{thread_rng, Rng};
+use serde::Serializer;
 use std::sync::atomic::{AtomicU64, Ordering};
 use uuid::Uuid;
 
@@ -41,13 +42,22 @@ pub(crate) fn float_to_string_for_hashing(x: f64) -> String {
     }
 }
 
-pub(crate) fn uuid_to_hex_string(uuid: Uuid) -> String {
-    let hex_string = uuid.as_bytes()
-        .iter()
-        .map(|byte| format!("{:02x}", byte))
-        .collect::<Vec<String>>()
-        .join("");
-    format!("0x{}", hex_string)
+pub(crate) fn as_hex_option<S>(cloid: &Option<Uuid>, s: S) -> std::result::Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    if let Some(cloid) = cloid {
+        s.serialize_str(&format!("0x{}", cloid.simple()))
+    } else {
+        s.serialize_none()
+    }
+}
+
+pub(crate) fn as_hex<S>(cloid: &Uuid, s: S) -> std::result::Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    s.serialize_str(&format!("0x{}", cloid.simple()))
 }
 
 pub(crate) fn generate_random_key() -> Result<[u8; 32]> {
@@ -100,8 +110,7 @@ impl BaseUrl {
 }
 
 lazy_static! {
-    static ref CUR_NONCE: AtomicU64 =
-        AtomicU64::new(now_timestamp_ms());
+    static ref CUR_NONCE: AtomicU64 = AtomicU64::new(now_timestamp_ms());
 }
 
 #[cfg(test)]
@@ -110,21 +119,42 @@ mod tests {
 
     #[test]
     fn float_to_string_for_hashing_test() {
-        assert_eq!(float_to_string_for_hashing(0.), "0".to_string()); 
+        assert_eq!(float_to_string_for_hashing(0.), "0".to_string());
         assert_eq!(float_to_string_for_hashing(-0.), "0".to_string());
         assert_eq!(float_to_string_for_hashing(-0.0000), "0".to_string());
-        assert_eq!(float_to_string_for_hashing(0.00076000), "0.00076".to_string());
-        assert_eq!(float_to_string_for_hashing(0.00000001), "0.00000001".to_string());
-        assert_eq!(float_to_string_for_hashing(0.12345678), "0.12345678".to_string());
+        assert_eq!(
+            float_to_string_for_hashing(0.00076000),
+            "0.00076".to_string()
+        );
+        assert_eq!(
+            float_to_string_for_hashing(0.00000001),
+            "0.00000001".to_string()
+        );
+        assert_eq!(
+            float_to_string_for_hashing(0.12345678),
+            "0.12345678".to_string()
+        );
         assert_eq!(
             float_to_string_for_hashing(87654321.12345678),
             "87654321.12345678".to_string()
         );
-        assert_eq!(float_to_string_for_hashing(987654321.00000000), "987654321".to_string());
-        assert_eq!(float_to_string_for_hashing(87654321.1234), "87654321.1234".to_string());
+        assert_eq!(
+            float_to_string_for_hashing(987654321.00000000),
+            "987654321".to_string()
+        );
+        assert_eq!(
+            float_to_string_for_hashing(87654321.1234),
+            "87654321.1234".to_string()
+        );
         assert_eq!(float_to_string_for_hashing(0.000760), "0.00076".to_string());
         assert_eq!(float_to_string_for_hashing(0.00076), "0.00076".to_string());
-        assert_eq!(float_to_string_for_hashing(987654321.0), "987654321".to_string());
-        assert_eq!(float_to_string_for_hashing(987654321.), "987654321".to_string());
+        assert_eq!(
+            float_to_string_for_hashing(987654321.0),
+            "987654321".to_string()
+        );
+        assert_eq!(
+            float_to_string_for_hashing(987654321.),
+            "987654321".to_string()
+        );
     }
 }